+++ XFGoogleMapSample/XFGoogleMapSample/CameraPage.xaml.cs
@@ -23,10 +23,36 @@ namespace XFGoogleMapSample
             map.Pins.Add(pinParis);
             map.Pins.Add(pinTokyo);
 
+            // CameraChanged is obsolete, please use CameraIdled
             map.CameraChanged += (sender, e) => 
             {
                 var p = e.Position;
-                labelStatus.Text = $"Lat={p.Target.Latitude:0.00}, Long={p.Target.Longitude:0.00}, Zoom={p.Zoom:0.00}, Bearing={p.Bearing:0.00}, Tilt={p.Tilt:0.00}";
+                var text = $"CameraChanged:Lat={p.Target.Latitude:0.00}, Long={p.Target.Longitude:0.00}, Zoom={p.Zoom:0.00}, Bearing={p.Bearing:0.00}, Tilt={p.Tilt:0.00}";
+                labelStatus.Text = text;
+                Debug.WriteLine(text);
+            };
+
+            map.CameraMoveStarted += (sender, e) => 
+            {
+                var text = $"CameraMoveStarted:IsGesture={e.IsGesture}";
+                labelStatus.Text = text;
+                Debug.WriteLine(text);
+            };
+
+            map.CameraMoving += (sender, e) =>
+            {
+                var p = e.Position;
+                var text = $"CameraMoving:Lat={p.Target.Latitude:0.00}, Long={p.Target.Longitude:0.00}, Zoom={p.Zoom:0.00}, Bearing={p.Bearing:0.00}, Tilt={p.Tilt:0.00}";
+                labelStatus.Text = text;
+                Debug.WriteLine(text);
+            };
+
+            map.CameraIdled += (sender, e) => 
+            {
+                var p = e.Position;
+                var text = $"CameraIdled:Lat={p.Target.Latitude:0.00}, Long={p.Target.Longitude:0.00}, Zoom={p.Zoom:0.00}, Bearing={p.Bearing:0.00}, Tilt={p.Tilt:0.00}";
+                labelStatus.Text = text;
+                Debug.WriteLine(text);
             };
 
             // MoveToCamera with Position