+++ Xamarin.Forms.GoogleMaps/Xamarin.Forms.GoogleMaps.Android/Logics/UiSettingsLogic.cs
@@ -6,10 +6,17 @@ namespace Xamarin.Forms.GoogleMaps.Logics.Android
 {
     internal sealed class UiSettingsLogic : BaseUiSettingsLogic<GoogleMap>
     {
+        // These properties are 'null' when call Initialize()
+        // If 'not null' then set true/false in app's page constructor.
+        public bool? RotateGesturesEnabled { get; private set; }
+        public bool? MyLocationButtonEnabled { get; private set; }
+        public bool? ScrollGesturesEnabled { get; private set; }
+        public bool? ZoomControlsEnabled { get; private set; }
+        public bool? ZoomGesturesEnabled { get; private set; }
+
         public override void Initialize()
         {
             base.Initialize();
-            NativeMap.UiSettings.MapToolbarEnabled = false;
         }
 
         protected override void OnUpdateCompassEnabled()
@@ -20,11 +27,13 @@ namespace Xamarin.Forms.GoogleMaps.Logics.Android
         protected override void OnUpdateRotateGesturesEnabled()
         {
             NativeMap.UiSettings.RotateGesturesEnabled = Map.UiSettings.RotateGesturesEnabled;
+            RotateGesturesEnabled = Map.UiSettings.RotateGesturesEnabled;
         }
 
         protected override void OnUpdateMyLocationButtonEnabled()
         {
             NativeMap.UiSettings.MyLocationButtonEnabled = Map.UiSettings.MyLocationButtonEnabled;
+            MyLocationButtonEnabled = Map.UiSettings.MyLocationButtonEnabled;
         }
 
         protected override void OnUpdateIndoorLevelPickerEnabled()
@@ -35,6 +44,7 @@ namespace Xamarin.Forms.GoogleMaps.Logics.Android
         protected override void OnUpdateScrollGesturesEnabled()
         {
             NativeMap.UiSettings.ScrollGesturesEnabled = Map.UiSettings.ScrollGesturesEnabled;
+            ScrollGesturesEnabled = Map.UiSettings.ScrollGesturesEnabled;
         }
 
         protected override void OnUpdateTiltGesturesEnabled()
@@ -45,11 +55,18 @@ namespace Xamarin.Forms.GoogleMaps.Logics.Android
         protected override void OnUpdateZoomControlsEnabled()
         {
             NativeMap.UiSettings.ZoomControlsEnabled = Map.UiSettings.ZoomControlsEnabled;
+            ZoomControlsEnabled = Map.UiSettings.ZoomControlsEnabled;
         }
 
         protected override void OnUpdateZoomGesturesEnabled()
         {
             NativeMap.UiSettings.ZoomGesturesEnabled = Map.UiSettings.ZoomGesturesEnabled;
+            ZoomGesturesEnabled = Map.UiSettings.ZoomGesturesEnabled;
+        }
+
+        protected override void OnUpdateMapToolbarEnabled()
+        {
+            NativeMap.UiSettings.MapToolbarEnabled = Map.UiSettings.MapToolbarEnabled;
         }
     }
 }