+++ Xamarin.Forms.GoogleMaps/Xamarin.Forms.GoogleMaps.Android/MapRenderer.cs
@@ -7,7 +7,6 @@ using Android.OS;
 using Java.Lang;
 using Xamarin.Forms.Platform.Android;
 using Math = System.Math;
-using Android.Util;
 using Android.App;
 using Android.Graphics;
 using Xamarin.Forms.GoogleMaps.Logics.Android;
@@ -17,30 +16,29 @@ using Android.Widget;
 using Android.Views;
 using Xamarin.Forms.GoogleMaps.Android.Logics;
 using Xamarin.Forms.GoogleMaps.Internals;
-using GCameraUpdateFactory = Android.Gms.Maps.CameraUpdateFactory;
-using GCameraPosition = Android.Gms.Maps.Model.CameraPosition;
 
 namespace Xamarin.Forms.GoogleMaps.Android
 {
-    public class MapRenderer : ViewRenderer,
-        GoogleMap.IOnCameraChangeListener,
+    public class MapRenderer : ViewRenderer<Map, global::Android.Views.View>,
         GoogleMap.IOnMapClickListener,
         GoogleMap.IOnMapLongClickListener,
         GoogleMap.IOnMyLocationButtonClickListener
     {
-        readonly CameraLogic _cameraLogic = new CameraLogic();
+        readonly CameraLogic _cameraLogic;
         readonly UiSettingsLogic _uiSettingsLogic = new UiSettingsLogic();
         readonly BaseLogic<GoogleMap>[] _logics;
 
         public MapRenderer() : base()
         {
+            _cameraLogic = new CameraLogic(UpdateVisibleRegion);
+
             AutoPackage = false;
             _logics = new BaseLogic<GoogleMap>[]
             {
                 new PolylineLogic(),
                 new PolygonLogic(),
                 new CircleLogic(),
-                new PinLogic(),
+                new PinLogic(OnMarkerCreating, OnMarkerCreated, OnMarkerDeleting, OnMarkerDeleted),
                 new TileLayerLogic(),
                 new GroundOverlayLogic()
             };
@@ -53,24 +51,26 @@ namespace Xamarin.Forms.GoogleMaps.Android
 
         protected GoogleMap NativeMap { get; private set; }
 
-        protected Map Map => (Map)Element;
+        protected Map Map => Element;
 
-        private GoogleMap _oldNativeMap;
-        private Map _oldMap;
+        private bool _ready = false;
+        private bool _onLayout = false;
 
-        bool _ready = false;
-        bool _onLayout = false;
-
-        float _scaledDensity;
+        private float _scaledDensity;
 
         public override SizeRequest GetDesiredSize(int widthConstraint, int heightConstraint)
         {
             return new SizeRequest(new Size(Context.ToPixels(40), Context.ToPixels(40)));
         }
 
-        protected override async void OnElementChanged(ElementChangedEventArgs<View> e)
+        protected override async void OnElementChanged(ElementChangedEventArgs<Map> e)
         {
             base.OnElementChanged(e);
+
+            if (e.OldElement == e.NewElement)
+            {
+                return;
+            }
 
             // For XAML Previewer or FormsGoogleMaps.Init not called.
             if (!FormsGoogleMaps.IsInitialized)
@@ -86,7 +86,30 @@ namespace Xamarin.Forms.GoogleMaps.Android
                 return;
             }
 
-            var oldMapView = (MapView)Control;
+            // Uninitialize old view
+            GoogleMap oldNativeMap = null;
+            Map oldMap = null;
+            if (e.OldElement != null)
+            {
+                try
+                {
+                    var oldNativeView = Control as MapView;
+                    oldNativeMap = await oldNativeView?.GetGoogleMapAsync();
+                    oldMap = e.OldElement;
+                    Uninitialize(oldNativeMap, oldMap);
+                    oldNativeView?.Dispose();
+                } 
+                catch (System.Exception ex) 
+                {
+                    var message = ex.Message;
+                    System.Diagnostics.Debug.WriteLine($"Uninitialize old view failed. - {message}");
+                }
+            }
+
+            if (e.NewElement == null)
+            {
+                return;
+            }
 
             var mapView = new MapView(Context);
             mapView.OnCreate(s_bundle);
@@ -96,52 +119,24 @@ namespace Xamarin.Forms.GoogleMaps.Android
             var activity = Context as Activity;
             if (activity != null)
             {
-                var metrics = new DisplayMetrics();
-                activity.WindowManager.DefaultDisplay.GetMetrics(metrics);
-                _scaledDensity = metrics.ScaledDensity;
+                _scaledDensity = activity.GetScaledDensity();
+                _cameraLogic.ScaledDensity = _scaledDensity;
                 foreach (var logic in _logics)
                 {
                     logic.ScaledDensity = _scaledDensity;
                 }
-
-                _cameraLogic.ScaledDensity = _scaledDensity;
             }
 
-            if (e.OldElement != null)
+            var newMap = e.NewElement;
+            NativeMap = await mapView.GetGoogleMapAsync();
+
+            foreach (var logic in _logics)
             {
-                var oldMapModel = (Map)e.OldElement;
-
-                oldMapModel.OnSnapshot -= OnSnapshot;
-                _cameraLogic.Unregister();
-
-                var oldGoogleMap = await oldMapView.GetGoogleMapAsync();
-                if (oldGoogleMap != null)
-                {
-
-                    oldGoogleMap.SetOnCameraChangeListener(null);
-                    oldGoogleMap.SetOnMapClickListener(null);
-                    oldGoogleMap.SetOnMapLongClickListener(null);
-                    oldGoogleMap.SetOnMyLocationButtonClickListener(null);
-                }
-
-                oldMapView.Dispose();
+                logic.Register(oldNativeMap, oldMap, NativeMap, newMap);
+                logic.ScaledDensity = _scaledDensity;
             }
 
-            if (oldMapView != null)
-            {
-                _oldNativeMap = await oldMapView.GetGoogleMapAsync();
-                _oldMap = (Map)e.OldElement;
-            }
-
-            NativeMap = await ((MapView)Control).GetGoogleMapAsync();
-
-            if (Map == null) return;
-            _cameraLogic.Register(Map, NativeMap);
-            Map.OnSnapshot += OnSnapshot;
-
-            _uiSettingsLogic.Register(Map, NativeMap);
-
-            OnMapReady(NativeMap);
+            OnMapReady(NativeMap, newMap);
         }
 
         private void OnSnapshot(TakeSnapshotMessage snapshotMessage)
@@ -155,31 +150,31 @@ namespace Xamarin.Forms.GoogleMaps.Android
             }));
         }
 
-        void OnMapReady(GoogleMap map)
+        private void OnMapReady(GoogleMap nativeMap, Map map)
         {
-            if (map != null)
+            if (nativeMap != null)
             {
-                map.SetOnCameraChangeListener(this);
-                map.SetOnMapClickListener(this);
-                map.SetOnMapLongClickListener(this);
-                map.SetOnMyLocationButtonClickListener(this);
+                _cameraLogic.Register(map, nativeMap);
 
-                _uiSettingsLogic.Initialize();
-                UpdateIsShowingUser();
-                UpdateHasScrollEnabled();
-                UpdateHasZoomEnabled();
-                UpdateHasRotationEnabled();
+                _uiSettingsLogic.Register(map, nativeMap);
+                Map.OnSnapshot += OnSnapshot;
+
+                nativeMap.SetOnMapClickListener(this);
+                nativeMap.SetOnMapLongClickListener(this);
+                nativeMap.SetOnMyLocationButtonClickListener(this);
+
+                UpdateIsShowingUser(_uiSettingsLogic.MyLocationButtonEnabled);
+                UpdateHasScrollEnabled(_uiSettingsLogic.ScrollGesturesEnabled);
+                UpdateHasZoomEnabled(_uiSettingsLogic.ZoomControlsEnabled, _uiSettingsLogic.ZoomGesturesEnabled);
+                UpdateHasRotationEnabled(_uiSettingsLogic.RotateGesturesEnabled);
                 UpdateIsTrafficEnabled();
                 UpdateIndoorEnabled();
                 UpdateMapStyle();
+                UpdateMyLocationEnabled();
+                _uiSettingsLogic.Initialize();
 
                 SetMapType();
                 SetPadding();
-            }
-
-            foreach (var logic in _logics)
-            {
-                logic.Register(_oldNativeMap, _oldMap, NativeMap, Map);
             }
 
             _ready = true;
@@ -215,7 +210,6 @@ namespace Xamarin.Forms.GoogleMaps.Android
         void InitializeLogic()
         {
             _cameraLogic.MoveCamera(Map.InitialCameraUpdate);
-            //_cameraLogic.MoveToRegion(((Map)Element).LastMoveToRegion, false);
 
             foreach (var logic in _logics)
             {
@@ -296,9 +290,10 @@ namespace Xamarin.Forms.GoogleMaps.Android
             }
         }
 
-        private void UpdateIsShowingUser()
+        private void UpdateIsShowingUser(bool? initialMyLocationButtonEnabled = null)
         {
-            NativeMap.MyLocationEnabled = NativeMap.UiSettings.MyLocationButtonEnabled = Map.IsShowingUser;
+            NativeMap.MyLocationEnabled = Map.IsShowingUser;
+            NativeMap.UiSettings.MyLocationButtonEnabled = initialMyLocationButtonEnabled ?? Map.IsShowingUser;
         }
 
         private void UpdateMyLocationEnabled()
@@ -306,20 +301,22 @@ namespace Xamarin.Forms.GoogleMaps.Android
             NativeMap.MyLocationEnabled = Map.MyLocationEnabled;
         }
 
-        private void UpdateHasScrollEnabled()
+        private void UpdateHasScrollEnabled(bool? initialScrollGesturesEnabled = null)
         {
-            NativeMap.UiSettings.ScrollGesturesEnabled = Map.HasScrollEnabled;
+            NativeMap.UiSettings.ScrollGesturesEnabled = initialScrollGesturesEnabled ?? Map.HasScrollEnabled;
         }
 
-        private void UpdateHasZoomEnabled()
+        private void UpdateHasZoomEnabled(
+            bool? initialZoomControlsEnabled = null, 
+            bool? initialZoomGesturesEnabled = null)
         {
-            NativeMap.UiSettings.ZoomControlsEnabled = Map.HasZoomEnabled;
-            NativeMap.UiSettings.ZoomGesturesEnabled = Map.HasZoomEnabled;
+            NativeMap.UiSettings.ZoomControlsEnabled = initialZoomControlsEnabled ?? Map.HasZoomEnabled;
+            NativeMap.UiSettings.ZoomGesturesEnabled = initialZoomGesturesEnabled ?? Map.HasZoomEnabled;
         }
 
-        private void UpdateHasRotationEnabled()
+        private void UpdateHasRotationEnabled(bool? initialRotateGesturesEnabled = null)
         {
-            NativeMap.UiSettings.RotateGesturesEnabled = Map.HasRotationEnabled;
+            NativeMap.UiSettings.RotateGesturesEnabled = initialRotateGesturesEnabled ?? Map.HasRotationEnabled;
         }
 
         private void UpdateIsTrafficEnabled()
@@ -376,14 +373,6 @@ namespace Xamarin.Forms.GoogleMaps.Android
                 (int)(Map.Padding.Bottom * _scaledDensity));
         }
 
-        public void OnCameraChange(GCameraPosition pos)
-        {
-            UpdateVisibleRegion(pos.Target);
-            var camera = pos.ToXamarinForms();
-            Map.CameraPosition = camera;
-            Map.SendCameraChanged(camera);
-        }
-
         public void OnMapClick(LatLng point)
         {
             Map.SendMapClicked(point.ToPosition());
@@ -423,31 +412,97 @@ namespace Xamarin.Forms.GoogleMaps.Android
             );
         }
 
+        #region Overridable Members
+
+        /// <summary>
+        /// Call when before marker create.
+        /// You can override your custom renderer for customize marker.
+        /// </summary>
+        /// <param name="outerItem">the pin.</param>
+        /// <param name="innerItem">the marker options.</param>
+        protected virtual void OnMarkerCreating(Pin outerItem, MarkerOptions innerItem)
+        {
+        }
+
+        /// <summary>
+        /// Call when after marker create.
+        /// You can override your custom renderer for customize marker.
+        /// </summary>
+        /// <param name="outerItem">the pin.</param>
+        /// <param name="innerItem">thr marker.</param>
+        protected virtual void OnMarkerCreated(Pin outerItem, Marker innerItem)
+        {
+        }
+
+        /// <summary>
+        /// Call when before marker delete.
+        /// You can override your custom renderer for customize marker.
+        /// </summary>
+        /// <param name="outerItem">the pin.</param>
+        /// <param name="innerItem">thr marker.</param>
+        protected virtual void OnMarkerDeleting(Pin outerItem, Marker innerItem)
+        {
+        }
+
+        /// <summary>
+        /// Call when after marker delete.
+        /// You can override your custom renderer for customize marker.
+        /// </summary>
+        /// <param name="outerItem">the pin.</param>
+        /// <param name="innerItem">thr marker.</param>
+        protected virtual void OnMarkerDeleted(Pin outerItem, Marker innerItem)
+        {
+        }
+
+        #endregion
+
+        private void Uninitialize(GoogleMap nativeMap, Map map) 
+        {
+            try
+            {
+                if (nativeMap == null)
+                {
+                    System.Diagnostics.Debug.WriteLine($"Uninitialize failed - {nameof(nativeMap)} is null");
+                    return;
+                }
+
+                if (map == null)
+                {
+                    System.Diagnostics.Debug.WriteLine($"Uninitialize failed - {nameof(map)} is null");
+                    return;
+                }
+
+                _uiSettingsLogic.Unregister();
+
+                map.OnSnapshot -= OnSnapshot;
+                _cameraLogic.Unregister();
+
+                foreach (var logic in _logics)
+                {
+                    logic.Unregister(nativeMap, map);
+                }
+
+                nativeMap.SetOnMapClickListener(null);
+                nativeMap.SetOnMapLongClickListener(null);
+                nativeMap.SetOnMyLocationButtonClickListener(null);
+
+                nativeMap.MyLocationEnabled = false;
+                nativeMap.Dispose();
+            }
+            catch (System.Exception ex)
+            {
+                var message = ex.Message;
+                System.Diagnostics.Debug.WriteLine($"Uninitialize failed. - {message}");
+            }
+        }
+
         bool _disposed;
         protected override void Dispose(bool disposing)
         {
             if (disposing && !_disposed)
             {
                 _disposed = true;
-
-                _uiSettingsLogic.Unregister();
-
-                Map.OnSnapshot -= OnSnapshot;
-                _cameraLogic.Unregister();
-
-                foreach (var logic in _logics)
-                    logic.Unregister(NativeMap, Map);
-
-                if (NativeMap != null)
-                {
-                    NativeMap.SetOnCameraChangeListener(null);
-                    NativeMap.SetOnMapClickListener(null);
-                    NativeMap.SetOnMapLongClickListener(null);
-                    NativeMap.SetOnMyLocationButtonClickListener(null);
-
-                    NativeMap.MyLocationEnabled = false;
-                    NativeMap.Dispose();
-                }
+                Uninitialize(NativeMap, Map);
             }
 
             base.Dispose(disposing);