+++ Xamarin.Forms.GoogleMaps/Xamarin.Forms.GoogleMaps.Android/Properties/GobalAssemblyInfo.cs
@@ -1,8 +1,9 @@
 ﻿using System.Reflection;
+using Xamarin.Forms.GoogleMaps.Internals;
 
-[assembly: AssemblyCompany("amay077")]
-[assembly: AssemblyProduct("Xamarin.Forms.GoogleMaps")]
-[assembly: AssemblyCopyright("Copyright © amay077. 2016 - 2017")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyVersion("2.1.1.0")]
-[assembly: AssemblyFileVersion("2.1.1.0")]
+[assembly: AssemblyCompany(ProductInformation.Author)]
+[assembly: AssemblyProduct(ProductInformation.Name)]
+[assembly: AssemblyCopyright(ProductInformation.Copyright)]
+[assembly: AssemblyTrademark(ProductInformation.Trademark)]
+[assembly: AssemblyVersion(ProductInformation.Version)]
+[assembly: AssemblyFileVersion(ProductInformation.Version)]