+++ Xamarin.Forms.GoogleMaps/Xamarin.Forms.GoogleMaps.iOS/Extensions/BitmapDescriptorExtensions.cs
@@ -2,6 +2,8 @@
 using UIKit;
 using Xamarin.Forms.Platform.iOS;
 using Foundation;
+using CoreImage;
+
 namespace Xamarin.Forms.GoogleMaps.iOS.Extensions
 {
     internal static class BitmapDescriptorExtensions
@@ -16,7 +18,8 @@ namespace Xamarin.Forms.GoogleMaps.iOS.Extensions
                     return UIImage.FromBundle(self.BundleName);
                 case BitmapDescriptorType.Stream:
                     self.Stream.Position = 0;
-                    return UIImage.LoadFromData(NSData.FromStream(self.Stream));
+                    // Resize to screen scale
+                    return UIImage.LoadFromData(NSData.FromStream(self.Stream), UIScreen.MainScreen.Scale);
                 case BitmapDescriptorType.AbsolutePath:
                     return UIImage.FromFile(self.AbsolutePath);
                 default: