+++ Xamarin.Forms.GoogleMaps/Xamarin.Forms.GoogleMaps/Properties/GlobalAssemblyInfo.cs
@@ -1,8 +1,0 @@
-﻿using System.Reflection;
-
-[assembly: AssemblyCompany("amay077")]
-[assembly: AssemblyProduct("Xamarin.Forms.GoogleMaps")]
-[assembly: AssemblyCopyright("Copyright © amay077. 2016 - 2017")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyVersion("2.1.1.0")]
-[assembly: AssemblyFileVersion("2.1.1.0")]